Move constant `NORMALIZATION` to `shared/` place
* where is shared with rule and alert processors.
