+++ stream_alert/shared/__init__.py
@@ -6,6 +6,7 @@ import os
 ALERT_PROCESSOR_NAME = 'alert_processor'
 ATHENA_PARTITION_REFRESH_NAME = 'athena_partition_refresh'
 RULE_PROCESSOR_NAME = 'rule_processor'
+NORMALIZATION_KEY = 'streamalert:normalization'
 
 # Create a package level logger to import
 LEVEL = os.environ.get('LOGGER_LEVEL', 'INFO').upper()
